Add game version and remove previously used states
Originally I was gonna have a menu and start cinematic, hence
the states. However as the deadline creeped in I realized what
a fool I was, so I had no choice but to scrap them. On the final
source code distribution on itch.io the files are gone as so are
the references. However that source distribution is slightly
different that this in term of asset paths too so I kinda want to
start anew with the original source code.